requirementxt.txt
amqp==5.0.6
appdirs==1.4.3
asgiref==3.3.1
billiard==3.6.4.0
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-import-export==2.5.0
django-seed==0.3.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
Faker==10.0.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
kombu==5.0.2
lockfile==0.12.2
MarkupPy==1.14
msgpack==0.6.2
mysqlclient==2.0.3
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.3
pandas==1.3.5
pep517==0.8.2
progress==1.5
prompt-toolkit==3.0.18
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
toposort==1.7
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
